wasmのソース分割
